feat: update package dependencies and add migration for managed_events table
- Updated dependencies in frontend package.json to specific versions for consistency.
- Added a new migration script for managed_events to ensure compatibility with legacy databases.
- Introduced MetricsMixin class for internal API metrics handling, providing various metrics retrieval methods.
