+++ requirements-dev.txt
@@ -1,5 +1,5 @@
 # Development dependencies
-pytest>=8.2.0
+pytest>=9.0.3
 pytest-asyncio>=0.23.6
 pytest-cov>=4.0.0
 pytest-timeout>=2.3.1
