+++ requirements.txt
@@ -1,7 +1,7 @@
 # Production dependencies
 discord.py>=2.3.1,<3.0.0
 PyYAML>=6.0,<7.0.0
-python-dotenv>=1.0.0,<2.0.0
+python-dotenv>=1.2.2,<2.0.0
 aiohttp>=3.13.3,<4.0.0
 beautifulsoup4>=4.12.2,<5.0.0
 lxml>=6.1.1,<7.0.0
@@ -16,6 +16,6 @@ httpx>=0.27.0,<1.0.0
 urllib3>=2.6.3,<3.0.0
 
 # Security pins for known CVE remediations
-cryptography>=46.0.5,<47.0.0
+cryptography>=48.0.2,<49.0.0
 pyasn1>=0.6.2,<0.7.0
 filelock>=3.20.3,<4.0.0
