+++ web/backend/requirements.txt
@@ -1,4 +1,4 @@
-fastapi>=0.128.0,<1.0.0
+fastapi>=0.137.1,<1.0.0
 uvicorn[standard]~=0.32.0
 python-multipart~=0.0.22
 httpx~=0.27.0
@@ -12,6 +12,6 @@ starlette>=1.3.1,<2.0.0
 beautifulsoup4>=4.12.2,<5.0.0
 lxml>=6.1.0,<7.0.0
 pyyaml~=6.0.0
-python-dotenv~=1.0.0
-pytest~=8.3.0
+python-dotenv>=1.2.2,<2.0.0
+pytest>=9.0.3
 pytest-asyncio~=0.24.0
