+++ web/frontend/package.json
@@ -13,33 +13,33 @@
     "format:check": "prettier --check src"
   },
   "dependencies": {
-    "axios": "^1.17.0",
-    "clsx": "^2.1.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hot-toast": "^2.6.0",
-    "react-router-dom": "^6.30.3",
-    "recharts": "^3.7.0"
+    "axios": "1.18.0",
+    "clsx": "2.1.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-hot-toast": "2.6.0",
+    "react-router-dom": "7.18.0",
+    "recharts": "3.7.0"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.1.0",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
-    "@vitejs/plugin-react": "^4.3.4",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^8.57.1",
-    "eslint-plugin-react": "^7.37.2",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "happy-dom": "^20.0.11",
-    "jsdom": "^27.3.0",
-    "postcss": "^8.5.15",
-    "prettier": "^3.3.3",
-    "tailwindcss": "^3.4.17",
-    "typescript": "^5.7.2",
-    "vite": "^6.0.3",
-    "vitest": "^4.1.8"
+    "@testing-library/jest-dom": "6.6.3",
+    "@testing-library/react": "16.1.0",
+    "@types/react": "18.3.12",
+    "@types/react-dom": "18.3.1",
+    "@typescript-eslint/eslint-plugin": "7.18.0",
+    "@typescript-eslint/parser": "7.18.0",
+    "@vitejs/plugin-react": "4.3.4",
+    "autoprefixer": "10.4.20",
+    "eslint": "8.57.1",
+    "eslint-plugin-react": "7.37.2",
+    "eslint-plugin-react-hooks": "4.6.2",
+    "happy-dom": "20.10.5",
+    "jsdom": "27.3.0",
+    "postcss": "8.5.15",
+    "prettier": "3.3.3",
+    "tailwindcss": "3.4.17",
+    "typescript": "5.7.2",
+    "vite": "6.4.3",
+    "vitest": "4.1.9"
   }
 }
